wrapper read_gbq function.
