+++ sdks/python/apache_beam/dataframe/transforms_test.py
@@ -18,27 +18,18 @@ import typing
 import unittest
 import warnings
 
-import mock
 import pandas as pd
 
 import apache_beam as beam
-import apache_beam.io.gcp.bigquery
 from apache_beam import coders
 from apache_beam import metrics
 from apache_beam.dataframe import convert
 from apache_beam.dataframe import expressions
 from apache_beam.dataframe import frame_base
 from apache_beam.dataframe import transforms
-from apache_beam.io.gcp.bigquery_tools import BigQueryWrapper
-from apache_beam.io.gcp.internal.clients import bigquery
 from apache_beam.runners.portability.fn_api_runner import fn_runner
 from apache_beam.testing.util import assert_that
 from apache_beam.testing.util import equal_to
-
-try:
-  from apitools.base.py.exceptions import HttpError
-except ImportError:
-  HttpError = None
 
 
 def check_correct(expected, actual):
@@ -432,58 +423,6 @@ class TransformPartsTest(unittest.TestCase):
           | 'SumB' >> beam.CombineGlobally(sum),
           equal_to([sB.sum()]),
           label='CheckValuesB')
-
-
-@unittest.skipIf(HttpError is None, 'GCP dependencies are not installed')
-class ReadGbqTransformTests(unittest.TestCase):
-  def test_without_specifying_table(self):
-    with self.assertRaisesRegex(
-        ValueError, "A table must be specified to Read from BigQuery."):
-      p = beam.Pipeline()
-      pipeline = p | beam.dataframe.transforms.ReadGbq()
-      pipeline
-
-  @mock.patch.object(BigQueryWrapper, 'get_table')
-  def test_bad_schema_public_api_direct_read(self, get_table):
-    fields = [
-        bigquery.TableFieldSchema(name='stn', type='DOUBLE', mode="NULLABLE"),
-        bigquery.TableFieldSchema(name='temp', type='FLOAT64', mode="REPEATED"),
-        bigquery.TableFieldSchema(name='count', type='INTEGER', mode=None)
-    ]
-    schema = bigquery.TableSchema(fields=fields)
-    table = apache_beam.io.gcp.internal.clients.bigquery. \
-        bigquery_v2_messages.Table(
-        schema=schema)
-    get_table.return_value = table
-
-    with self.assertRaisesRegex(ValueError,
-                                "Encountered an unsupported type: 'DOUBLE'"):
-      p = apache_beam.Pipeline()
-      pipeline = p | apache_beam.dataframe.transforms.ReadGbq(
-          table="dataset.sample_table", use_bqstorage_api=True)
-      pipeline
-
-  def test_unsupported_callable(self):
-    def filterTable(table):
-      if table is not None:
-        return table
-
-    res = filterTable
-    with self.assertRaisesRegex(TypeError,
-                                'ReadFromBigQuery: table must be of type string'
-                                '; got a callable instead'):
-      p = beam.Pipeline()
-      pipeline = p | beam.dataframe.transforms.ReadGbq(table=res)
-      pipeline
-
-  def test_ReadGbq_unsupported_param(self):
-    with self.assertRaisesRegex(ValueError,
-                                "Unsupported parameter entered in ReadGbq. "
-                                "Please enter only supported parameters."):
-      p = beam.Pipeline()
-      pipeline = p | beam.dataframe.transforms.ReadGbq(
-          table="table", reauth="true_config")
-      pipeline
 
 
 if __name__ == '__main__':
